\encoding{UTF-8}
\name{diazinon}
\docType{data}
\alias{diazinon}
\title{GUTS data set with survivors exposed to diazinon}
\description{
3 experiments with survivors exposed to concentrations of diazinon. The list contains 12 variables.
}
\usage{diazinon}
\format{A list containing 12 vectors, plus a description string.}
%\source{<none>}
\references{Jager T., Albert C., Preuss T., Ashauer R., General unified theory of survival – a toxicokinetic toxicodynamic framework for ecotoxicology, published online in Env. Sci. and Tech. (2011) (\href{http://www.mendeley.com/research/general-unified-threshold-model-survival-toxicokinetictoxicodynamic-framework-ecotoxicology/\#}{http://www.mendeley.com/research/general-unified-threshold-model-survival-toxicokinetictoxicodynamic-framework-ecotoxicology/#}).}
\keyword{datasets}

\author{%
Carlo Albert & Sören Vogel \email{guts.package@gmail.com}

Maintainer: Carlo Albert & Sören Vogel \email{guts.package@gmail.com}%
}

\seealso{\code{\link{GUTS}}}
