\encoding{UTF-8}
\name{diazinon}
\alias{diazinon}
\docType{data}
\title{GUTS data set with Gammarus pulex exposed to diazinon}
\description{Data of 3 pulsed toxicity tests with the freshwater crustacean Gammarus pulex and diazinon, an organophosphate insecticide.}
\usage{data("diazinon")}
\format{A list containing 12 data vectors.}
\source{Ashauer, R., Hintermeister, A., Caravatti, I., Kretschmann, A. and Escher, B. I. (2010) Toxicokinetic-toxicodynamic modeling explains carry-over toxicity from exposure to diazinon by slow organism recovery. \emph{Environmental Science & Technology} 44, 3963--3971, \cr
\href{http://pubs.acs.org/doi/abs/10.1021/es903478b}{http://pubs.acs.org/doi/abs/10.1021/es903478b}.}
\keyword{datasets}
\author{Carlo Albert \email{carlo.albert@eawag.ch}, Sören Vogel \email{soeren.vogel@posteo.ch}, Oliver Jakoby \email{oliver.jakoby@rifcon.de}, Alexander Singer \email{alexander.singer@rifcon.de} and Dirk Nickisch \email{dirk.nickisch@rifcon.de}
Maintainer: Oliver Jakoby \email{oliver.jakoby@rifcon.de}}
\seealso{\code{\link{GUTS}}}
