\encoding{UTF-8}
\name{diazinon}
\docType{data}
\alias{diazinon}
\title{GUTS data set with survivors exposed to diazinon}
\description{3 experiments with survivors exposed to concentrations of diazinon. The list contains 12 variables.}
\usage{diazinon}
\format{A list containing 12 vectors, plus a description string.}
\source{Ashauer, R., Hintermeister, A., Caravatti, I., Kretschmann, A. & Escher, B. I. 2010 Toxicokinetic-toxicodynamic modeling explains carry-over toxicity from exposure to diazinon by slow organism recovery. Environmental Science & Technology 44, 3963-3971, \href{http://pubs.acs.org/doi/abs/10.1021/es903478b}{http://pubs.acs.org/doi/abs/10.1021/es903478b}.}
%\references{}
\keyword{datasets}

\author{Carlo Albert \email{carlo.albert@eawag.ch} & Sören Vogel \email{soeren.vogel@posteo.ch}

Maintainer: Sören Vogel \email{soeren.vogel@posteo.ch}}

\seealso{\code{\link{GUTS}}}
